Changing overriding location with current location
When loading the script into Scriptable, it would only allow me to view a location in Boston and would curiously not ask me for my location.

I figured that if I forced to ask for my location by changing the override location from Boston to my current location, I could get it to work.

I am sure this is not the best way to implement it, but maybe it helps fixing the script for future users.