Adding one more row in table to lead to Special:Nearby
While using the widget I found that sometimes there are too many things of interest nearby and the one I want to check out is not there. Hence, I would have found it useful to have one more item in the bottom that leads me to the Wikipedia Nearby page. Hence, I added it in.